Added HeightArrayFactory Interface in order to implement different World Generators

+++ src/main/java/org/eann/sim/simulation/Tile.java
@@ -29,17 +29,17 @@
 package org.eann.sim.simulation;
 
 public class Tile {
-    private float hight;
+    private double hight;
     private int x;
     private int y;
 
-    public Tile(float hight, int x, int y) {
+    public Tile(double hight, int x, int y) {
         this.hight = hight;
         this.x = x;
         this.y = y;
     }
 
-    public float getHeight() {
+    public double getHeight() {
         return this.hight;
     }
